chg: :recycle: :truck: make a digraph.py module
